+++ src/cdo/cs_equation.c
@@ -1068,6 +1068,7 @@ cs_equation_add(const char            *eqname,
   case CS_EQUATION_TYPE_PREDEFINED:
   case CS_EQUATION_TYPE_GROUNDWATER:
   case CS_EQUATION_TYPE_MAXWELL:
+  case CS_EQUATION_TYPE_THERMAL:
   case CS_EQUATION_TYPE_NAVSTO:
     _n_predef_equations++;
     break;