+++ src/cdo/cs_equation_param.h
@@ -138,6 +138,9 @@ BEGIN_C_DECLS
  * Equation related to the resolution of the Navier-Stokes system
  * - Example: momentum, prediction, correction, energy...
  *
+ * \var CS_EQUATION_TYPE_THERMAL
+ * Equation related to the heat transfer
+ *
  * \var CS_EQUATION_TYPE_PREDEFINED
  * Predefined equation (most part of the setting is already done)
  * - Example: equation for the wall distance or ALE
@@ -149,6 +152,7 @@ typedef enum {
   CS_EQUATION_TYPE_GROUNDWATER,
   CS_EQUATION_TYPE_MAXWELL,
   CS_EQUATION_TYPE_NAVSTO,
+  CS_EQUATION_TYPE_THERMAL,
   CS_EQUATION_TYPE_PREDEFINED,
   CS_EQUATION_N_TYPES